docker/resnet/app/requirements.txt
tensorflow-cpu==2.2.0
keras==2.3.1
opencv-python==4.2.0.34
flask==1.1.2
numpy==1.17.4
scipy==1.4.1